Address review feedback for XXE vulnerability
