+++ src/main/java/org/sasanlabs/service/vulnerability/xxe/TimeoutInputStream.java
@@ -1,0 +1,43 @@
+package org.sasanlabs.service.vulnerability.xxe;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+
+class TimeoutInputStream extends InputStream {
+    private final InputStream delegate;
+    private final ScheduledFuture<?> future;
+
+    public TimeoutInputStream(
+            InputStream delegate, long timeoutMs, ScheduledThreadPoolExecutor scheduler) {
+        this.delegate = delegate;
+        this.future =
+                scheduler.schedule(
+                        () -> {
+                            try {
+                                delegate.close();
+                            } catch (IOException e) {
+                                // Ignore
+                            }
+                        },
+                        timeoutMs,
+                        java.util.concurrent.TimeUnit.MILLISECONDS);
+    }
+
+    @Override
+    public int read() throws IOException {
+        return delegate.read();
+    }
+
+    @Override
+    public int read(byte[] b, int off, int len) throws IOException {
+        return delegate.read(b, off, len);
+    }
+
+    @Override
+    public void close() throws IOException {
+        future.cancel(true);
+        delegate.close();
+    }
+}
