+++ src/main/java/org/sasanlabs/service/vulnerability/xxe/TimingXMLFilter.java
@@ -1,0 +1,48 @@
+package org.sasanlabs.service.vulnerability.xxe;
+
+import org.xml.sax.Attributes;
+import org.xml.sax.SAXException;
+import org.xml.sax.XMLReader;
+import org.xml.sax.helpers.XMLFilterImpl;
+
+class TimingXMLFilter extends XMLFilterImpl {
+    private final long startTime;
+    private final long timeoutMs;
+
+    public TimingXMLFilter(XMLReader parent, long timeoutMs) {
+        super(parent);
+        this.startTime = System.currentTimeMillis();
+        this.timeoutMs = timeoutMs;
+    }
+
+    private void checkTimeout() throws SAXException {
+        if (System.currentTimeMillis() - startTime > timeoutMs) {
+            throw new SAXException("XML parsing cooperative timeout exceeded");
+        }
+    }
+
+    @Override
+    public void startElement(String uri, String localName, String qName, Attributes attributes)
+            throws SAXException {
+        checkTimeout();
+        super.startElement(uri, localName, qName, attributes);
+    }
+
+    @Override
+    public void characters(char[] ch, int start, int length) throws SAXException {
+        checkTimeout();
+        super.characters(ch, start, length);
+    }
+
+    @Override
+    public void ignorableWhitespace(char[] ch, int start, int length) throws SAXException {
+        checkTimeout();
+        super.ignorableWhitespace(ch, start, length);
+    }
+
+    @Override
+    public void processingInstruction(String target, String data) throws SAXException {
+        checkTimeout();
+        super.processingInstruction(target, data);
+    }
+}
