+++ src/main/java/org/sasanlabs/service/vulnerability/xxe/XXEVulnerability.java
@@ -1,8 +1,6 @@
 package org.sasanlabs.service.vulnerability.xxe;
 
-import java.io.IOException;
 import java.io.InputStream;
-import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.Semaphore;
 import javax.servlet.http.HttpServletRequest;
@@ -19,6 +17,8 @@ import org.apache.logging.log4j.Logger;
 import org.sasanlabs.internal.utility.LevelConstants;
 import org.sasanlabs.internal.utility.Variant;
 import org.sasanlabs.internal.utility.annotations.AttackVector;
+import org.sasanlabs.internal.utility.annotations.ChallengeCard;
+import org.sasanlabs.internal.utility.annotations.ChallengeCard.Payload;
 import org.sasanlabs.internal.utility.annotations.VulnerableAppRequestMapping;
 import org.sasanlabs.internal.utility.annotations.VulnerableAppRestController;
 import org.sasanlabs.service.vulnerability.bean.GenericVulnerabilityResponseBean;
@@ -31,11 +31,9 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.xml.sax.Attributes;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
-import org.xml.sax.helpers.XMLFilterImpl;
 
 /**
  * Resources referred while writing this Vulnerability. <br>
@@ -67,7 +65,7 @@ public class XXEVulnerability {
     private static ScheduledThreadPoolExecutor createTimeoutScheduler() {
         ScheduledThreadPoolExecutor executor =
                 new ScheduledThreadPoolExecutor(
-                        1,
+                        2,
                         runnable -> {
                             Thread thread = new Thread(runnable);
                             thread.setDaemon(true);
@@ -296,6 +294,12 @@ public class XXEVulnerability {
             value = LevelConstants.LEVEL_6,
             htmlTemplate = "LEVEL_1/XXE",
             requestMethod = RequestMethod.POST)
+    @ChallengeCard(
+            challengeText = "XXE_BILLION_LAUGHS_CHALLENGE",
+            payload =
+                    @Payload(
+                            description = "XXE_BILLION_LAUGHS_PAYLOAD_DESC",
+                            value = "XXE_BILLION_LAUGHS_PAYLOAD_VAL"))
     public ResponseEntity<GenericVulnerabilityResponseBean<Book>> getVulnerablePayloadLevel6(
             HttpServletRequest request) {
         if (!billionLaughsSemaphore.tryAcquire()) {
@@ -348,84 +352,5 @@ public class XXEVulnerability {
         }
         return new ResponseEntity<>(
                 new GenericVulnerabilityResponseBean<>(null, false), HttpStatus.OK);
-    }
-
-    static class TimeoutInputStream extends InputStream {
-        private final InputStream delegate;
-        private final ScheduledFuture<?> future;
-
-        public TimeoutInputStream(
-                InputStream delegate, long timeoutMs, ScheduledThreadPoolExecutor scheduler) {
-            this.delegate = delegate;
-            this.future =
-                    scheduler.schedule(
-                            () -> {
-                                try {
-                                    delegate.close();
-                                } catch (IOException e) {
-                                    // Ignore
-                                }
-                            },
-                            timeoutMs,
-                            java.util.concurrent.TimeUnit.MILLISECONDS);
-        }
-
-        @Override
-        public int read() throws IOException {
-            return delegate.read();
-        }
-
-        @Override
-        public int read(byte[] b, int off, int len) throws IOException {
-            return delegate.read(b, off, len);
-        }
-
-        @Override
-        public void close() throws IOException {
-            future.cancel(true);
-            delegate.close();
-        }
-    }
-
-    static class TimingXMLFilter extends XMLFilterImpl {
-        private final long startTime;
-        private final long timeoutMs;
-
-        public TimingXMLFilter(XMLReader parent, long timeoutMs) {
-            super(parent);
-            this.startTime = System.currentTimeMillis();
-            this.timeoutMs = timeoutMs;
-        }
-
-        private void checkTimeout() throws SAXException {
-            if (System.currentTimeMillis() - startTime > timeoutMs) {
-                throw new SAXException("XML parsing cooperative timeout exceeded");
-            }
-        }
-
-        @Override
-        public void startElement(String uri, String localName, String qName, Attributes attributes)
-                throws SAXException {
-            checkTimeout();
-            super.startElement(uri, localName, qName, attributes);
-        }
-
-        @Override
-        public void characters(char[] ch, int start, int length) throws SAXException {
-            checkTimeout();
-            super.characters(ch, start, length);
-        }
-
-        @Override
-        public void ignorableWhitespace(char[] ch, int start, int length) throws SAXException {
-            checkTimeout();
-            super.ignorableWhitespace(ch, start, length);
-        }
-
-        @Override
-        public void processingInstruction(String target, String data) throws SAXException {
-            checkTimeout();
-            super.processingInstruction(target, data);
-        }
     }
 }
