+++ src/test/java/org/sasanlabs/service/vulnerability/xxe/XXEVulnerabilityTest.java
@@ -422,7 +422,7 @@ public class XXEVulnerabilityTest {
         SAXParserFactory spf = SAXParserFactory.newInstance();
         XMLReader parent = spf.newSAXParser().getXMLReader();
 
-        XXEVulnerability.TimingXMLFilter filter = new XXEVulnerability.TimingXMLFilter(parent, 0);
+        TimingXMLFilter filter = new TimingXMLFilter(parent, 0);
         Thread.sleep(1);
 
         org.assertj.core.api.Assertions.assertThatThrownBy(
@@ -468,8 +468,8 @@ public class XXEVulnerabilityTest {
                         }
                     };
 
-            XXEVulnerability.TimeoutInputStream timeoutStream =
-                    new XXEVulnerability.TimeoutInputStream(blockingStream, 100, scheduler);
+            TimeoutInputStream timeoutStream =
+                    new TimeoutInputStream(blockingStream, 100, scheduler);
 
             long startTime = System.currentTimeMillis();
             org.assertj.core.api.Assertions.assertThatThrownBy(
@@ -535,8 +535,7 @@ public class XXEVulnerabilityTest {
         try {
             byte[] data = "Hello World".getBytes();
             InputStream innerStream = new ByteArrayInputStream(data);
-            XXEVulnerability.TimeoutInputStream timeoutStream =
-                    new XXEVulnerability.TimeoutInputStream(innerStream, 2000, scheduler);
+            TimeoutInputStream timeoutStream = new TimeoutInputStream(innerStream, 2000, scheduler);
 
             byte[] buf = new byte[5];
             int readBytes = timeoutStream.read(buf, 0, 5);
@@ -552,15 +551,13 @@ public class XXEVulnerabilityTest {
     public void testTimingXMLFilter_IgnorableWhitespaceAndProcessingInstruction() throws Exception {
         SAXParserFactory spf = SAXParserFactory.newInstance();
         XMLReader parent = spf.newSAXParser().getXMLReader();
-        XXEVulnerability.TimingXMLFilter filter =
-                new XXEVulnerability.TimingXMLFilter(parent, 2000);
+        TimingXMLFilter filter = new TimingXMLFilter(parent, 2000);
 
         filter.ignorableWhitespace(new char[0], 0, 0);
         filter.processingInstruction("target", "data");
         filter.characters(new char[0], 0, 0);
 
-        XXEVulnerability.TimingXMLFilter timeoutFilter =
-                new XXEVulnerability.TimingXMLFilter(parent, 0);
+        TimingXMLFilter timeoutFilter = new TimingXMLFilter(parent, 0);
         Thread.sleep(1);
 
         org.assertj.core.api.Assertions.assertThatThrownBy(
